Generified `DataReader` and moved to LBJava-NLP-tools package Added missing license header to `PrepSRLAnnotatorTest` Added high-level Javadoc comments
